+++ src/Credfeto.Dispatcher.Discord/Services/DiscordWebhookDispatcher.cs
@@ -7,6 +7,8 @@ using System.Threading.Tasks;
 using Credfeto.Dispatcher.Discord.Configuration;
 using Credfeto.Dispatcher.Discord.DataTypes;
 using Credfeto.Dispatcher.Discord.Interfaces;
+using Credfeto.Dispatcher.Discord.Services.LoggingExtensions;
+using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 
 namespace Credfeto.Dispatcher.Discord.Services;
@@ -14,12 +16,14 @@ namespace Credfeto.Dispatcher.Discord.Services;
 public sealed class DiscordWebhookDispatcher : IDiscordDispatcher
 {
     private readonly HttpClient _httpClient;
+    private readonly ILogger<DiscordWebhookDispatcher> _logger;
     private readonly DiscordOptions _options;
 
-    public DiscordWebhookDispatcher(HttpClient httpClient, IOptions<DiscordOptions> options)
+    public DiscordWebhookDispatcher(HttpClient httpClient, IOptions<DiscordOptions> options, ILogger<DiscordWebhookDispatcher> logger)
     {
         this._httpClient = httpClient;
         this._options = options.Value;
+        this._logger = logger;
     }
 
     public async ValueTask SendAsync(DiscordMessage message, CancellationToken cancellationToken)
@@ -40,6 +44,12 @@ public sealed class DiscordWebhookDispatcher : IDiscordDispatcher
             using HttpRequestMessage request = new(method: HttpMethod.Post, requestUri: this._options.WebhookUrl);
             request.Content = new StringContent(content: json, encoding: Encoding.UTF8, mediaType: "application/json");
             response = await this._httpClient.SendAsync(request: request, cancellationToken: cancellationToken);
+
+            if (!response.IsSuccessStatusCode)
+            {
+                this._logger.LogWebhookNonSuccess(statusCode: (int)response.StatusCode);
+            }
+
             _ = response.EnsureSuccessStatusCode();
         }
         finally